Feature/jenkins (#9)
* Add Jenkinsfiles for building and pushing Docker images

Added Jenkinsfiles for the cart-service, eureka-server, gateway-service, order-service and product-service. Each file handles Maven packaging, testing, Docker building and finally pushing the Docker images to the Docker Hub, including a separate command to tag and push images on the main branch.

* Add Helm main stage in Jenkinsfile

A Helm main stage was added to the Jenkinsfile in multiple directories. Also, adjustments were made to add the short Git commit to the Docker tag for better versioning during push to the Docker Hub. Paths in gateway-service and front-end YAML files were also duplicated for ImplementationSpecific pathType.